requirements.txt
selenium
PyYAML
beautifulsoup4